💚 [fix] only bring necessary testing dependencies
bw2data and matrix_utils are necessary, not the whole brightway2[5] package.
Previous pipeline setup was installing brightway2, hence bringing all
the deps of the brightway2 metapackage, but the master branch of the
bw2cal repo is now for brightway25, so we shouldn't bring old deps, and
we should only bring the necessary ones (not bw2io for example)

This commit also adds a pip virtual environment for the failning
ubuntu-pip setup of the testing

👷 test code being checked-out by installing first

By installing the code from source, we test the current master branch, or the code offered in a PR.

⬇️ package has dependency on bw2data not brightway2

Using bw2data as dependency keeps a coherent set of deps.
Testing dependencies are only about third party libraries required for testing, and are different
to *install_requires*.